Fix Windows compilation error

+++ src/runtime_src/xdp/profile/database/events/creator/device_event_from_trace.cpp
@@ -273,7 +273,7 @@ namespace xdp {
 
   void DeviceEventCreatorFromTrace::end()
   {
-    for(uint64_t amIndex = 0; amIndex < cuStarts.size(); ++amIndex) {
+    for(uint32_t amIndex = 0; amIndex < cuStarts.size(); ++amIndex) {
       if(cuStarts[amIndex].empty()) {
         continue;
       }
